requirements.txt
discord.py==2.4.0
ccxt==4.3.36
pandas==2.2.2
numpy==1.26.4
aiohttp==3.9.5
requests==2.32.3
tabulate==0.9.0
